requirements.txt
google-auth==2.38.0
google-genai==1.5.0
requests==2.32.3
